feat: add tabs to map and node popups for better iOS scrolling (#1705)
Add tabbed interface to node popups to fix iOS rubber-banding when
scrolling long traceroute data. The tabs use icon-only buttons to
preserve space:
- Info tab (ℹ️): Node details and DM button
- Traceroute tab (📡): Traceroute data and traceroute button

Tabs only appear when user has traceroute permissions.

- Create MapNodePopupContent component for Leaflet map popups
- Update NodesTab to use new component, clean up unused imports
- Add tab support to NodePopup component for consistency
- Add CSS styles for tabs in both App.css and NodePopup.css
- Add translation keys for tab labels
