+++ src/components/NodesTab.tsx
@@ -7,11 +7,10 @@ import { TabType } from '../types/ui';
 import { ResourceType } from '../types/permission';
 import { createNodeIcon, getHopColor } from '../utils/mapIcons';
 import { generateArrowMarkers } from '../utils/mapHelpers.tsx';
-import { getEffectivePosition, getHardwareModelName, getRoleName, hasValidEffectivePosition, isNodeComplete, parseNodeId, TRACEROUTE_DISPLAY_HOURS } from '../utils/nodeHelpers';
-import { formatTime, formatDateTime, formatRelativeTime } from '../utils/datetime';
+import { getEffectivePosition, getRoleName, hasValidEffectivePosition, isNodeComplete, parseNodeId } from '../utils/nodeHelpers';
+import { formatTime, formatDateTime } from '../utils/datetime';
 import { getDistanceToNode } from '../utils/distance';
 import { getTilesetById } from '../config/tilesets';
-import { formatTracerouteRoute } from '../utils/traceroute';
 import { getEffectiveHops } from '../utils/nodeHops';
 import { useMapContext } from '../contexts/MapContext';
 import { useTelemetryNodes, useDeviceConfig, useNodes } from '../hooks/useServerData';
@@ -30,6 +29,7 @@ import PacketMonitorPanel from './PacketMonitorPanel';
 import { getPacketStats } from '../services/packetApi';
 import { NodeFilterPopup } from './NodeFilterPopup';
 import { VectorTileLayer } from './VectorTileLayer';
+import { MapNodePopupContent } from './MapNodePopupContent';
 
 /**
  * Spiderfier initialization constants
@@ -1547,180 +1547,21 @@ const NodesTabComponent: React.FC<NodesTabProps> = ({
                   </Tooltip>
                 )}
                 <Popup autoPan={false}>
-                  <div className="node-popup">
-                    <div className="node-popup-header">
-                      <div className="node-popup-title">{node.user?.longName || `Node ${node.nodeNum}`}</div>
-                      {node.user?.shortName && (
-                        <div className="node-popup-subtitle">{node.user.shortName}</div>
-                      )}
-                    </div>
-
-                    <div className="node-popup-grid">
-                      {node.user?.id && (
-                        <div className="node-popup-item">
-                          <span className="node-popup-icon">🆔</span>
-                          <span className="node-popup-value">{node.user.id}</span>
-                        </div>
-                      )}
-
-                      {node.user?.role !== undefined && (() => {
-                        const roleNum = typeof node.user.role === 'string'
-                          ? parseInt(node.user.role, 10)
-                          : node.user.role;
-                        const roleName = getRoleName(roleNum);
-                        return roleName ? (
-                          <div className="node-popup-item">
-                            <span className="node-popup-icon">👤</span>
-                            <span className="node-popup-value">{roleName}</span>
-                          </div>
-                        ) : null;
-                      })()}
-
-                      {node.user?.hwModel !== undefined && (() => {
-                        const hwModelName = getHardwareModelName(node.user.hwModel);
-                        return hwModelName ? (
-                          <div className="node-popup-item">
-                            <span className="node-popup-icon">🖥️</span>
-                            <span className="node-popup-value">{hwModelName}</span>
-                          </div>
-                        ) : null;
-                      })()}
-
-                      {node.snr != null && (
-                        <div className="node-popup-item">
-                          <span className="node-popup-icon">📶</span>
-                          <span className="node-popup-value">{node.snr.toFixed(1)} dB</span>
-                        </div>
-                      )}
-
-                      {(() => {
-                        const popupHops = getEffectiveHops(node, nodeHopsCalculation, traceroutes, currentNodeNum);
-                        return popupHops < 999 ? (
-                          <div className="node-popup-item">
-                            <span className="node-popup-icon">🔗</span>
-                            <span className="node-popup-value">{popupHops} hop{popupHops !== 1 ? 's' : ''}</span>
-                          </div>
-                        ) : null;
-                      })()}
-
-                      {node.position?.altitude != null && (
-                        <div className="node-popup-item">
-                          <span className="node-popup-icon">⛰️</span>
-                          <span className="node-popup-value">{node.position.altitude}m</span>
-                        </div>
-                      )}
-
-                      {node.deviceMetrics?.batteryLevel !== undefined && node.deviceMetrics.batteryLevel !== null && (
-                        <div className="node-popup-item">
-                          <span className="node-popup-icon">{node.deviceMetrics.batteryLevel === 101 ? '🔌' : '🔋'}</span>
-                          <span className="node-popup-value">
-                            {node.deviceMetrics.batteryLevel === 101 ? 'Plugged In' : `${node.deviceMetrics.batteryLevel}%`}
-                          </span>
-                        </div>
-                      )}
-                    </div>
-
-                    {node.lastHeard && (
-                      <div className="node-popup-footer">
-                        <span className="node-popup-icon">🕐</span>
-                        {formatDateTime(new Date(node.lastHeard * 1000), timeFormat, dateFormat)}
-                      </div>
-                    )}
-
-                    {/* Traceroute Section */}
-                    {(() => {
-                      if (!currentNodeId || !node.user?.id || currentNodeId === node.user.id) return null;
-
-                      // Get current node number from ID
-                      const currentNodeNum = parseNodeId(currentNodeId);
-                      if (currentNodeNum === null) return null;
-
-                      // Use shared constant for traceroute visibility window
-                      const cutoff = Date.now() - TRACEROUTE_DISPLAY_HOURS * 60 * 60 * 1000;
-
-                      // Find the most recent traceroute between these two nodes
-                      const recentTraceroute = traceroutes
-                        .filter(tr => {
-                          const isRelevant =
-                            (tr.fromNodeNum === currentNodeNum && tr.toNodeNum === node.nodeNum) ||
-                            (tr.fromNodeNum === node.nodeNum && tr.toNodeNum === currentNodeNum);
-                          return isRelevant && tr.timestamp >= cutoff;
-                        })
-                        .sort((a, b) => b.timestamp - a.timestamp)[0];
-
-                      if (!recentTraceroute) return null;
-
-                      return (
-                        <div className="node-popup-traceroute">
-                          <div className="traceroute-header">
-                            <strong>{t('node_popup.last_traceroute')}</strong>
-                            <span className="traceroute-age">
-                              ({formatRelativeTime(recentTraceroute.timestamp)})
-                            </span>
-                          </div>
-                          {recentTraceroute.route && recentTraceroute.route !== 'null' ? (
-                            <>
-                              <div className="traceroute-path">
-                                <span className="traceroute-label">{t('node_popup.forward_path')}:</span>
-                                <span className="traceroute-route">
-                                  {formatTracerouteRoute(
-                                    recentTraceroute.route,
-                                    recentTraceroute.snrTowards,
-                                    recentTraceroute.fromNodeNum,
-                                    recentTraceroute.toNodeNum,
-                                    nodes,
-                                    distanceUnit
-                                  )}
-                                </span>
-                              </div>
-                              {recentTraceroute.routeBack && recentTraceroute.routeBack !== 'null' && (
-                                <div className="traceroute-path">
-                                  <span className="traceroute-label">{t('node_popup.return_path')}:</span>
-                                  <span className="traceroute-route">
-                                    {formatTracerouteRoute(
-                                      recentTraceroute.routeBack,
-                                      recentTraceroute.snrBack,
-                                      recentTraceroute.toNodeNum,
-                                      recentTraceroute.fromNodeNum,
-                                      nodes,
-                                      distanceUnit
-                                    )}
-                                  </span>
-                                </div>
-                              )}
-                            </>
-                          ) : (
-                            <div className="traceroute-failed">
-                              {t('node_popup.traceroute_failed')}
-                            </div>
-                          )}
-                        </div>
-                      );
-                    })()}
-
-                    {node.user?.id && hasPermission('messages', 'read') && (
-                      <button
-                        className="node-popup-btn"
-                        onClick={handlePopupDMClick(node)}
-                      >
-                        💬 Direct Message
-                      </button>
-                    )}
-                    {node.user?.id && hasPermission('traceroute', 'write') && onTraceroute && (
-                      <button
-                        className="node-popup-btn"
-                        onClick={() => onTraceroute(node.user!.id)}
-                        disabled={connectionStatus !== 'connected' || tracerouteLoading === node.user?.id}
-                      >
-                        {tracerouteLoading === node.user?.id ? (
-                          <span className="spinner"></span>
-                        ) : (
-                          '📡'
-                        )}{' '}
-                        {t('node_popup.traceroute', 'Traceroute')}
-                      </button>
-                    )}
-                  </div>
+                  <MapNodePopupContent
+                    node={node}
+                    nodes={nodes}
+                    currentNodeId={currentNodeId}
+                    timeFormat={timeFormat}
+                    dateFormat={dateFormat}
+                    distanceUnit={distanceUnit}
+                    traceroutes={traceroutes}
+                    hasPermission={hasPermission}
+                    onDMNode={handlePopupDMClick(node)}
+                    onTraceroute={onTraceroute ? () => onTraceroute(node.user!.id) : undefined}
+                    connectionStatus={connectionStatus}
+                    tracerouteLoading={tracerouteLoading}
+                    getEffectiveHops={(n) => getEffectiveHops(n, nodeHopsCalculation, traceroutes, currentNodeNum)}
+                  />
                 </Popup>
               </Marker>
                 );
